+++ supervisor/confecho.py
@@ -1,8 +1,8 @@
 import sys
-from supervisor import resources
 from supervisor.compat import as_string
+from supervisor.compat import resource_file
 
 
 def main(out=sys.stdout):
-    config = resources.read_text(__package__, 'skel/sample.conf')
-    out.write(as_string(config))
+    with open(resource_file(__package__, 'skel/sample.conf'), 'r') as f:
+        out.write(as_string(f.read()))